Fix phpswitch script
When using phpswitch with empty or incorrect command, ir ended VM session instead of script (because this script is used with source command).
Also as phpversionformer is now trying to get correct version, it does not return 'off' as a version any more, so additional checks has to be done.