Use extra mounting options only for Windows
This is needed as not all platforms support and/or accept the given
parameters and sometimes one could get the following message which
would prevent from the VM to start at all, e.g.:

```
Failed to mount folders in Linux guest. You've specified mount options
which are not supported by "prl_fs" file system.

Invalid mount options: ["dmode=775,fmode=664"]
```

Also the original issue only seems to affect Windows platform with
virtualbox as hypervisor, as this can be seen at:

https://github.com/ansible/ansible/issues/42388